signature Added

+++ src/main/java/org/dvare/expression/operation/aggregation/Sum.java
@@ -1,18 +1,18 @@
 /**
  * The MIT License (MIT)
- *
+ * <p>
  * Copyright (c) 2016-2017 DVARE (Data Validation and Aggregation Rule Engine)
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Sogiftware.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -36,6 +36,10 @@ import org.dvare.expression.veriable.VariableExpression;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * @author Muhammad Hammad
+ * @since 2016-06-30
+ */
 @Operation(type = OperationType.SUM, dataTypes = {DataType.FloatType, DataType.IntegerType})
 public class Sum extends AggregationOperationExpression {
     static Logger logger = LoggerFactory.getLogger(Sum.class);
@@ -58,11 +62,11 @@ public class Sum extends AggregationOperationExpression {
             switch (type) {
 
                 case FloatType: {
-                    leftExpression = LiteralType.getLiteralExpression(Float.valueOf(0f), variableExpression.getType());
+                    leftExpression = LiteralType.getLiteralExpression(0f, variableExpression.getType());
                     break;
                 }
                 case IntegerType: {
-                    leftExpression = LiteralType.getLiteralExpression(Integer.valueOf(0), variableExpression.getType());
+                    leftExpression = LiteralType.getLiteralExpression(0, variableExpression.getType());
                     break;
                 }
                 case StringType: {
